safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFECompositeElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 118: TypeError: 'undefined' is not an object (evaluating 'overComposite1.in1.baseVal = "F1"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL overComposite1.in1.baseVal should be F12. Threw exception TypeError: 'undefined' is not an object (evaluating 'overComposite1.in1.baseVal')
